Refactor components of net into their own modules
