+++ components/builder-sessionsrv/src/lib.rs
@@ -18,8 +18,6 @@ extern crate habitat_core as hab_core;
 extern crate habitat_net as hab_net;
 extern crate hyper;
 #[macro_use]
-extern crate lazy_static;
-#[macro_use]
 extern crate log;
 extern crate num_cpus;
 extern crate protobuf;
