+++ components/builder-vault/Cargo.toml
@@ -11,7 +11,6 @@ doc = false
 
 [dependencies]
 env_logger = "*"
-lazy_static = "*"
 log = "*"
 num_cpus = "*"
 protobuf = "*"
