+++ components/builder-vault/src/lib.rs
@@ -17,8 +17,6 @@ extern crate habitat_builder_protocol as protocol;
 extern crate habitat_core as hab_core;
 extern crate habitat_net as hab_net;
 #[macro_use]
-extern crate lazy_static;
-#[macro_use]
 extern crate log;
 extern crate num_cpus;
 extern crate protobuf;
