+++ components/net/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Adam Jacob <adam@chef.io>", "Jamie Winsor <reset@chef.io>", "Fletche
 [dependencies]
 fnv = "*"
 hyper = "*"
+lazy_static = "*"
 libc = "*"
 log = "*"
 protobuf = "*"
